requirements.txt
Authlib==0.12.1
boto3==1.11.6
Click==7.0
connexion==2.5.1
Flask==1.1.1
jsonschema==2.6.0
passlib==1.7.2
pathlib==1.0.1
prettytable==0.7.2
prometheus-client==0.7.1
prometheus-flask-exporter==0.12.1
psutil==5.6.7
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-swiftclient==3.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
semantic-version==2.6.0
six==1.14.0
SQLAlchemy==1.3.12
swagger-spec-validator==2.4.3
toastedmarshmallow==2.15.2.post1
treelib==1.5.5
Twisted[tls]==20.3.0
typing==3.7.4.1
urllib3==1.25.8
watchdog==0.9.0
Werkzeug==0.16.0
yosai==0.3.2
zope.component==4.6
zope.interface==4.7.2
ijson==2.5.1
google-cloud-storage==1.31.0

# Optional and third-party/integration dependencies
python-keystoneclient==3.22.0
pg8000==1.10.6
